fix(ci): satisfy transcript lint checks
